Fixed bugs in memory layouts, generated getters for fields, added a type check to castFrom()
