+++ generator/src/main/java/io/github/jwharm/javagi/generator/Conversions.java
@@ -198,4 +198,19 @@ public class Conversions {
             default -> toCamelCase(primitive, true);
         };
     }
+
+    public static String literal(String type, String value) throws NumberFormatException {
+        return switch (type) {
+            case "boolean" -> Boolean.valueOf(value).toString();
+            case "byte" -> Byte.valueOf(value).toString();
+            case "char" -> "'" + value + "'";
+            case "double" -> Double.valueOf(value) + "d";
+            case "float" -> Float.valueOf(value) + "f";
+            case "int" -> Integer.valueOf(value).toString();
+            case "long" -> Long.valueOf(value) + "L";
+            case "short" -> Short.valueOf(value).toString();
+            case "java.lang.String" -> '"' + value + '"';
+            default -> value;
+        };
+    }
 }
