+++ generator/src/main/java/io/github/jwharm/javagi/model/Bitfield.java
@@ -17,6 +17,7 @@ public class Bitfield extends ValueWrapper {
 
         writer.write("public class " + javaName + " extends io.github.jwharm.javagi.Bitfield {\n");
         
+        generateCType(writer);
         generateMemoryLayout(writer);
         
         for (Member m : memberList) {
