+++ generator/src/main/java/io/github/jwharm/javagi/model/Class.java
@@ -6,9 +6,12 @@ import java.io.IOException;
 import java.io.Writer;
 
 public class Class extends RegisteredType {
+	
+	public String typeStruct;
 
-    public Class(GirElement parent, String name, String parentClass, String cType, String version) {
+    public Class(GirElement parent, String name, String parentClass, String cType, String typeStruct, String version) {
         super(parent, name, parentClass, cType, version);
+        this.typeStruct = typeStruct;
     }
 
     public void generate(Writer writer) throws IOException {
@@ -35,11 +38,16 @@ public class Class extends RegisteredType {
         writer.write(" {\n");
 
         generateEnsureInitialized(writer);
+        generateCType(writer);
         generateMemoryLayout(writer);
+        for (Field f : fieldList) {
+        	f.generate(writer);
+        }
+
         generateMemoryAddressConstructor(writer);
         generateCastFromGObject(writer);
         generateConstructors(writer);
-
+        
         for (Method m : methodList) {
             m.generate(writer, false, false);
         }
