+++ generator/src/main/java/io/github/jwharm/javagi/model/Constant.java
@@ -3,6 +3,8 @@ package io.github.jwharm.javagi.model;
 import java.io.IOException;
 import java.io.Writer;
 
+import io.github.jwharm.javagi.generator.Conversions;
+
 public class Constant extends GirElement {
 
     public final String value, cType;
@@ -26,30 +28,15 @@ public class Constant extends GirElement {
             String printValue;
             if (type.isAliasForPrimitive()) {
                 String aliasedType = ((Alias) type.girElementInstance).type.simpleJavaType;
-                printValue = "new " + type.qualifiedJavaType + "(" + literal(aliasedType, value) + ")";
+                printValue = "new " + type.qualifiedJavaType + "(" + Conversions.literal(aliasedType, value) + ")";
             } else if (type.isEnum()) {
-                printValue = type.qualifiedJavaType + ".fromValue(" + literal("int", value) + ")";
+                printValue = type.qualifiedJavaType + ".fromValue(" + Conversions.literal("int", value) + ")";
             } else {
-                printValue = literal(type.qualifiedJavaType, value);
+                printValue = Conversions.literal(type.qualifiedJavaType, value);
             }
             writer.write("    public static final " + type.qualifiedJavaType + " " + name + " = " + printValue + ";\n");
         } catch (NumberFormatException nfe) {
             // Do not write anything
         }
-    }
-
-    private String literal(String type, String value) throws NumberFormatException {
-        return switch (type) {
-            case "boolean" -> Boolean.valueOf(value).toString();
-            case "byte" -> Byte.valueOf(value).toString();
-            case "char" -> "'" + value + "'";
-            case "double" -> Double.valueOf(value) + "d";
-            case "float" -> Float.valueOf(value) + "f";
-            case "int" -> Integer.valueOf(value).toString();
-            case "long" -> Long.valueOf(value) + "L";
-            case "short" -> Short.valueOf(value).toString();
-            case "java.lang.String" -> '"' + value + '"';
-            default -> value;
-        };
     }
 }
